PageObject HW commit refactor
